==> 01_datawrangling/10_crimes.ipynb <==
# Data wrangling for crime data
1. Read data
2. Groupby time periods


import pandas as pd
import numpy as np
os.chdir('/Users/sarwaridas/Desktop/IDS 702/Final Project/ids720_CoffeeAndCrimes/')
#crime = pd.read_csv(os.getcwd()+"/00_sourcedata/Crimes_-_2001_to_Present.csv")
crime=pd.read_csv("/Users/sarwaridas/Desktop/IDS 702/Final Project/Crimes_-_2001_to_Present.csv")
#crime=crime.loc[crime['Community Area']!=0]


crime = crime.drop(
    [
        "Updated On",
        "Latitude",
        "Longitude",
        "FBI Code",
        "X Coordinate",
        "Y Coordinate",
        "Case Number",
        "Block",
        "IUCR",
        "Arrest",
        "Domestic",
        "Beat",
    ],
    axis=1,
)
crime.nunique() #77 community areas

crime1=crime.loc[crime.Year>2004]
crime1=crime1.loc[crime1.Year<2020]
crime1.shape

crime_by_year_CCA1 = crime1[["Year", "Community Area","ID"]].groupby(["Year", "Community Area"], as_index=False).count()
crime_by_year_CCA1.shape 

crime_by_year_CCA1

(100 * crime.isna().sum()) / len(crime)  # Missing 8.14% data for Community Area

crime.head()

crime_by_year_CCA = crime[["Year", "Community Area","ID"]].groupby(["Year", "Community Area"], as_index=False).count()
crime_by_year_CCA = crime_by_year_CCA.loc[crime_by_year_CCA["Year"] != 2001] #something is up with numbers for 2001
crime_by_year_CCA.sample(2) 

crime_by_year_CCA.Year.s

# import altair as alt
# alt.Chart(crime_by_year_CCA).mark_circle(size=60).encode(
#     x="Year:O", y="ID:Q", color="Ward:N", tooltip=["Community Area", "Year", "ID"]
# ).interactive()


# CCA = crime_by_year_CCA.Ward.unique()
# input_dropdown = alt.binding_select(options=CCA)
# selection = alt.selection_single(fields=["Community Area"], bind=input_dropdown, name="CCA:")
# color = alt.condition(
#     selection, alt.Color("Community Area:N", legend=None), alt.value("lightgray")
# )

# alt.Chart(crime_by_year_CCA).mark_point().encode(
#     x=alt.X("Year:O", scale=alt.Scale(zero=False)),
#     y=alt.Y("ID:Q", scale=alt.Scale(zero=False)),
#     color="Community Area:N",
#     tooltip=["Community Area", "Year", "ID"],
# ).add_selection(selection).transform_filter(selection)

os.chdir('/Users/sarwaridas/Desktop/IDS 702/Final Project/ids720_CoffeeAndCrimes/02_processed_data')
crime_by_year_CCA.to_csv("crime_by_year_CCA.csv")

crime_by_year_CCA.columns
