==> 01_datawrangling/20_coffee.ipynb <==
# Main predictor: Coffee shops
1. Read data
2. Extract licenses with coffee or cafe in them
3. Extract year they were licensed
4. Convert area to Chicago Community Area

import pandas as pd
import numpy as np
import os
os.chdir('/Users/sarwaridas/Desktop/IDS 702/Final Project/ids720_CoffeeAndCrimes/00_sourcedata')
col_list = [
    "ID",
    "LEGAL NAME",
    "DOING BUSINESS AS NAME",
    "ADDRESS",
    "CITY",
    "STATE",
    "ZIP CODE",
    "WARD",
    "PRECINCT",
    "WARD PRECINCT",
    "LICENSE STATUS",
    "LICENSE STATUS CHANGE DATE",
    "LICENSE DESCRIPTION",
    "BUSINESS ACTIVITY ID",
    "BUSINESS ACTIVITY",
    "LICENSE TERM START DATE",
    "LICENSE TERM EXPIRATION DATE",
    "DATE ISSUED",
    "LATITUDE",
    "LONGITUDE",
    'LOCATION'
]
df_full = pd.read_csv("/Users/sarwaridas/Desktop/Business_Licenses.csv",
    usecols=col_list,
    low_memory=False,
)

# https://data.cityofchicago.org/Community-Economic-Development/Business-Licenses/r5kz-chrr


## Cleaning for coffee shops and extracting year

df_full = df_full.loc[
    (df_full["STATE"] == "IL") & (df_full["CITY"] == "CHICAGO")
]  # subsetting for chicago
names_doingbusiness = [str(x).lower() for x in df_full["DOING BUSINESS AS NAME"]]
names_businessactivity = [str(x).lower() for x in df_full["BUSINESS ACTIVITY"]]
names_legalactivity = [str(x).lower() for x in df_full["LEGAL NAME"]]

keep_row = []
for a, b, c in zip(names_doingbusiness, names_businessactivity, names_legalactivity):
    keep = 0
    if (
        ("coffee" in a)
        or ("coffee" in b)
        or ("coffee" in c)
        or ("cafe" in a)
        or ("cafe" in b)
        or ("coffee" in c)
    ):
        keep = 1
    keep_row.append(keep)


df_full.insert(loc=0, column="names_doingbusiness", value=names_doingbusiness)
df_full.insert(loc=1, column="names_businessactivity", value=names_businessactivity)
df_full.insert(loc=2, column="names_legalactivity", value=names_legalactivity)
df_full.insert(loc=3, column="keep_row", value=keep_row)
df = df_full.loc[df_full.keep_row == 1]
df = df.drop(["LEGAL NAME", "DOING BUSINESS AS NAME", "CITY", "STATE"], axis=1)


df["lc_start_date"] = pd.to_datetime(df["LICENSE TERM START DATE"], format="%m/%d/%Y")
df['year']=df['lc_start_date'].map(lambda x: x.year)
#df.insert(loc=5, column='years', value=years)
df=df.loc[df.year>2000]
df = df.drop(
    [
        "LICENSE TERM START DATE",
        "LICENSE TERM EXPIRATION DATE",
        "DATE ISSUED",
        "LICENSE STATUS",
        "LICENSE STATUS CHANGE DATE",
    ],
    axis=1,
)
df.head()


df.drop(['names_businessactivity', 'names_legalactivity','ADDRESS', 'ZIP CODE', 'WARD', 'PRECINCT',
       'WARD PRECINCT', 'LICENSE DESCRIPTION', 'BUSINESS ACTIVITY ID',
       'BUSINESS ACTIVITY','lc_start_date'],axis=1,inplace=True)

df.to_csv("CoffeeShops_byLatitude.csv")

# Mapping to Community Areas

#### Option A: This code demonstrates how to take the shapefile provided on the City of Chicago Open Data Portal and converts it to a JSON file, with top-level keys named for each neighborhood, and associated values that give the latitude and longitude pairs corresponding to the neighborhood's boundaries. For example: (https://github.com/craigmbooth/chicago_neighborhood_finder)

#Fails to build GDAL

#### Option B: Determining Chicago neighborhood from latitude and longitude. The script uses a GeoJSON file containing polygons of Chicago neighborhoods. Once all of the polygons have been generated, a simple point in polygon algorithm is used to determine which neighborhood contains the point of interest. (https://github.com/jkgiesler/parse-chicago-neighborhoods/blob/master/gps_to_neighborhood.py)

df_loc=pd.DataFrame(df.LONGITUDE)
df_loc['LATITUDE']=pd.DataFrame(df.LATITUDE)

loc = pd.read_csv("Locations_withCCA.csv")
lookup= pd.read_csv("location_lookup.csv")

loc.CCA=loc.CCA.str.title()

loc=loc.replace({'Andersonville': 'Edgewater','Archer Heights,West Elsdon':'Archer Heights','Armour Square,Chinatown':'Armour Square',
'Belmont Cragin,Hermosa':'Belmont Cragin','Avalon Park,Calumet Heights':'Avalon Park','Belmont Craigin,Hermosa':'Belmont Cragin',
'Brighton Park,Mckinley Park':'Brighton Park','Wicker Park,West Town':'West Town','Back Of The Yards':'New City','Bronzeville':'Douglas','Bucktown':'Logan Square',
'Chatham,Burnside':'Burnside', 'Garfield Park': 'East Garfield Park','Wrigleyville':'Lakeview','West Loop':'Near West Side','Washington Heights,Roseland':'Washington Heights',
'United Center':'Near West Side','Ukrainian Village And East Village':'West Town','Streeterville': 'Near North Side', 'Southeast Side':'South Chicago',
'South Shore, Grand Crossing':'South Shore','Sheffield & Depaul':'Lincoln Park', 'Sauganash,Forest Glen': 'Forest Glen', 'Rush & Division':'Near North Side',  'River North': 'Near North Side',
 'Printers Row':'Loop', 'Old Town': 'Near North Side', 'North Park,Albany Park':'Albany Park', 'Museum Campus':'Loop', 'Mount Greenwood,Morgan Park': 'Morgan Park',
  'Gold Coast':'Lincoln Park',
 'Grant Park':'Loop','Greektown':'Near West Side', 'Irving Park,Avondale': 'Avondale', 'Kenwood,Oakland': 'Kenwood','Jackson Park':'South Deering','Little Italy, Uic':'Near West Side','Marquette Park,Gage Park':'Gage Park',
 'Montclare, Galewood':'Montclare','Midway Airport':'Garfield Ridge','Little Village':'South Lawndale','Lakeview':'Lake View','Millenium Park':'Loop'
})


# df_loc.to_csv("Locations.csv")
## run gps_to_neighborhood.py
locations = loc
locations['LOCATION'] = list(zip(locations.LATITUDE, locations.LONGITUDE))
locations = locations.drop(["Unnamed: 0", "Unnamed: 0.1"], axis=1)
locations.head()


locations['CCA']= np.where(locations.LATITUDE.isnull(),np.nan, locations['CCA'])

#(100 * locations.isna().sum()) / len(locations)  # Missing 8.14% data for Wards


lookup.columns=['CCA_id','CCA']

locations_new=pd.merge(locations,lookup,on="CCA",how="left")
locations_new.CCA.nunique()

locations_new.shape

locations.shape

df.shape

df['CCA']=pd.DataFrame(locations['CCA'].values)
df=df.reset_index()

df['CCA']=pd.DataFrame(locations['CCA'].values)

(100 * df.isna().sum()) / len(df)  # Missing 8.14% data for Wards


coffee_by_year_CCA=df[["year","CCA","keep_row"]].groupby(["CCA","year"],as_index=False).count()

import pandas as pd
os.chdir('/Users/sarwaridas/Desktop/IDS 702/Final Project/ids720_CoffeeAndCrimes/02_processed_data')
#coffee_by_year_CCA.to_csv("coffee_by_year_CCA.csv")
coffee_by_year_CCA=pd.read_csv("coffee_by_year_CCA.csv")

coffee_by_year_CCA.head()

coffee_by_year_CCA.CCA.nunique()

coffee_by_year_CCA.CCA.unique()
