==> 01_datawrangling/27_makefinaldf.ipynb <==
import pandas as pd
import numpy as np
import os
os.chdir('/Users/sarwaridas/Desktop/IDS 702/Final Project/ids720_CoffeeAndCrimes/')
coffee_old=pd.read_csv("02_processed_data/coffee_by_year_CCA.csv")
coffee=pd.read_csv("02_processed_data/coffee_by_year_CCA_revised.csv")
crime=pd.read_csv("02_processed_data/crime_by_year_CCA.csv")

coffee_old=coffee_old.rename(columns={'keep_row':'coffeeshops'})
coffee_old=coffee_old.drop(['Unnamed: 0'],axis=1)
coffee_old=coffee_old.pivot_table(values="coffeeshops", index="CCA", columns="year").stack(dropna=False).reset_index()
coffee_old.head()

coffee=coffee.drop(['Unnamed: 0'],axis=1)
coffee=coffee.rename(columns={'keep_row':'coffeeshops'})
coffee['community']=coffee['community'].apply(lambda x: x.title())
coffee.head()



# coffee=coffee.rename(columns={'keep_row':'coffeeshops'})
# coffee=coffee.drop(['Unnamed: 0'],axis=1)
# coffee=coffee.pivot_table(values="coffeeshops", index="CCA", columns="year").stack(dropna=False).reset_index()

lookup= pd.read_csv(os.getcwd()+"/02_processed_data/location_lookup.csv")
lookup= dict(zip(lookup.CCA, lookup.name))

crime['CCA_name']=crime["Community Area"].map(lookup)
crime['CCA_name']=crime['CCA_name'].apply(lambda x: x.title())

crime=crime.drop(['Unnamed: 0',"Community Area"],axis=1)
crime=crime.rename(columns={'ID':'crimes'})

crime.head()

crime.nunique()

crime=crime.pivot_table(values="crimes", index="CCA_name", columns="Year").stack(dropna=False).reset_index()
crime.columns=['CCA_name', 'Year', 'crimes']


crime.head()

coffee.head()


df_yearly=crime.merge(coffee, how='outer', left_on=["Year",'CCA_name'], right_on=["year","community"],validate="1:1")


# df_yearly['coffeeshops']=df_yearly['coffeeshops'].interpolate()
df_yearly=df_yearly.interpolate()
df_yearly.drop(['community','year'],inplace=True,axis=1)
df_yearly.head(34)


df_yearly.to_csv(os.getcwd()+"/02_processed_data/df_crimes_coffee_by_year.csv")

df_yearly.groupby(['CCA_name']).sum()

#coffee.columns=['CCA', 'year', 'coffeeshops', 'timeperiod']

# df_full=crime.merge(coffee, how='outer', left_on=["Year",'CCA_name'], right_on=["year","CCA"])
# df_full.head()

#df_full.sample(10)

df_yearly.isna().sum()/len(df_yearly)

df_yearly.sample(10)

df_yearly.head()

def category(x):
    if x <= 2009:
        return '2005-2009'
    elif x <= 2014: #& x>= 2010:
        return '2010-2014'
    elif x <= 2019:# & x>= 2015:
        return '2015-2019'
    elif x== 2020:
        return '2020'

# crime=crime.loc[crime.Year>=2005]
# crime=crime.loc[crime.Year<2021]

# coffee=coffee.loc[coffee.year>=2005]
# coffee=coffee.loc[coffee.year<2021]

# crime.Year=crime.Year.astype(int)
# coffee.year=coffee.year.astype(int)

df_yearly['timeperiod']=df_yearly.Year.apply(lambda x: category(x))
# coffee['timeperiod']=coffee.year.apply(lambda x: category(x))



df_by_timeperiod=df_yearly.groupby(['timeperiod','CCA_name']).sum().reset_index() #change this
df_by_timeperiod.drop("Year",inplace=True,axis=1)

df_by_timeperiod.sample(10)

df_by_timeperiod.to_csv(os.getcwd()+"/02_processed_data/coffee_crime_revised.csv")
