==> 01_datawrangling/30_controls.ipynb <==
The ACS launched in 2005 as a replacement to the Census Long Form. Unfortunately, the test run data from 2000-2004 is not geotagged beyond the state level.


import pandas as pd
import numpy as np
import os

os.chdir('/Users/sarwaridas/Desktop/IDS 702/Final Project/ids720_CoffeeAndCrimes')
df= pd.read_csv("02_processed_data/coffee_crime_revised.csv")

df.isna().sum()

df=df.drop(['Unnamed: 0'],axis=1)

df.head()

import statsmodels.api as sm
import statsmodels.formula.api as smf
smf.ols("crimes ~ coffeeshops+ C(timeperiod)+ C(CCA_name)", data=df).fit().summary()


#https://www.census.gov/programs-surveys/acs/guidance/comparing-acs-data/2019.html
#https://cookcountypublichealth.org/wp-content/uploads/2018/12/SES-FINAL.pdf

## Collecting Control variables

cols1= ["CCA","TOT_POP",'WHITE','BLACK','ASIAN',"HISP",'POP_HH','UNEMP','EMP', 'INC_LT_25K',
 'INC_25_50K',
 'INC_50_75K',
 'INC_75_100K',
 'INC_100_150K',
 'INC_GT_150', 'MEDINC','HS','BACH','MED_AGE']

cols2=["GEOG","TOT_POP",'WHITE','BLACK','ASIAN',"HISP",'POP_HH','UNEMP','EMP', 'INC_LT_25K',
 'INC_25_50K',
 'INC_50_75K',
 'INC_75_100K',
 'INC_100_150K',
 'INC_GT_150', 'MEDINC','HS','BACH','MED_AGE']

cca0509 = pd.read_csv(os.getcwd() + "/00_sourcedata/Reference_CCAProfiles_2005_2009.csv",usecols=cols1)
cca1014 = pd.read_csv(os.getcwd() + "/00_sourcedata/Reference_CCAProfiles_2010_2014.csv",usecols=cols2)
cca1519 = pd.read_csv(os.getcwd() + "/00_sourcedata/ReferenceCCAProfiles20152019.csv",usecols=cols2) #https://datahub.cmap.illinois.gov/dataset/1d2dd970-f0a6-4736-96a1-3caeb431f5e4/resource/d23fc5b1-0bb5-4bcc-bf70-688201534833/download/CDSFieldDescriptions202106.pdf


# cca0509.columns = [str(col) + '_0509' for col in cca0509.columns]
# cca1014.columns = [str(col) + '_1014' for col in cca1014.columns]
# cca1519.columns = [str(col) + '_1519' for col in cca1519.columns]
# cca1014.head()

df=df.loc[df.timeperiod!='2020']

df.timeperiod.value_counts()

cca0509['timeperiod']='2005-2009'
cca1014['timeperiod']='2010-2014'
cca1519['timeperiod']='2015-2019'


cca0509= cca0509.rename(columns={'CCA':'GEOG'})


cca0509.columns

cca1014.columns

controls=cca0509.append(cca1014)
controls=controls.append(cca1519)
controls.sample(10)

controls.isna().sum()

df.columns

#getting_there=pd.merge(df,cca1014,how="outer",left_on=['timeperiod','CCA_name'],right_on=["timeperiod","GEOG_1014_1014"])
df=df.rename(columns={'timeperiod_x':'timeperiod','CCA_name':'GEOG'})
df_byperiod= df.groupby(['timeperiod','GEOG'],as_index=False).mean()
df.head()

controls.head()

controls.groupby(['GEOG','timeperiod']).count()

df_merged= pd.merge(df_byperiod,controls,how="left",on=['GEOG','timeperiod'])

df_merged.head()

import statsmodels.api as sm
import statsmodels.formula.api as smf

smf.ols(
    "crimes ~ coffeeshops+ C(timeperiod) + GEOG + TOT_POP+ WHITE+HISP+ BLACK+ ASIAN+POP_HH+ EMP+UNEMP+INC_LT_25K+INC_25_50K+INC_50_75K+INC_75_100K+INC_100_150K+INC_GT_150+MEDINC",
    data=df_merged,
).fit().summary()


df_merged.columns

df_merged['AsianPerc']= df_merged.ASIAN / df_merged.TOT_POP
df_merged['BlackPerc']= df_merged.BLACK / df_merged.TOT_POP
df_merged['HispPerc']= df_merged.HISP / df_merged.TOT_POP
df_merged['WhitePerc']= df_merged.WHITE / df_merged.TOT_POP
df_merged['UnemployedPerc']= df_merged.UNEMP / df_merged.TOT_POP
df_merged['EmployedPerc']= df_merged.EMP / df_merged.TOT_POP
df_merged['Inc25k']= df_merged.INC_LT_25K / df_merged.TOT_POP
df_merged['Inc50k']= df_merged.INC_25_50K / df_merged.TOT_POP
df_merged['Inc75k']= df_merged.INC_50_75K / df_merged.TOT_POP
df_merged['Inc100k']= df_merged.INC_75_100K / df_merged.TOT_POP
df_merged['Inc150k']= df_merged.INC_100_150K / df_merged.TOT_POP
df_merged['Inc_gt150k']= df_merged.INC_GT_150 / df_merged.TOT_POP
df_merged['HSperc']= df_merged.HS / df_merged.TOT_POP
df_merged['BACHperc']= df_merged.BACH / df_merged.TOT_POP



df_merged.head()

os.getcwd()
os.chdir('/Users/sarwaridas/Desktop/IDS 702/Final Project/ids720_CoffeeAndCrimes/02_processed_data')

df_merged.to_csv("coffee_crime_controls_revised.csv")
